Extract StandalonePluginBuilder
Move the WITHOUT_BINDINGS rendering shape (a plain Plugin<Project> with
an empty apply() body and no binding annotation) into a dedicated
subclass. PluginClassBuilder.renderJava / renderKotlin delegate to it
when the plugin type is WITHOUT_BINDINGS, so existing call sites that
use PluginClassBuilder as a DSL collector keep working.
